requirements.txt
dj-database-url==0.5.0
Django==3.2.4
gensim==3.8.0
google-images-download==2.8.0
joblib==1.0.1
jupyter==1.0.0
kiwisolver==1.3.1
lmdb==1.2.1
matplotlib==3.3.4
nltk==3.6.2
numpy==1.17.0
Pillow==6.2.2
regex==2021.4.4
requests==2.25.1
scipy==1.5.4
selenium==3.141.0
simple-image-download==0.4
spacy==2.1.3
sqlparse==0.4.1
torch==1.8.1
torchvision==0.2.1
tqdm==4.61.0
whitenoise==5.2.0

